For deco stops show the gas of the next segment in the table. In recalculation remove old deco stops earlier.
In struct diveplan, the items are "segments" with a beginning, a duration, and a gas. In contrast, the UI of the planner uses
"waypoints" which are the boundaries between segments. It is conventional at least for reco stops to display the gas of the _next_
segment in the runtime table (i.e. the gas possibly to be switched to).

Furthermore, in addStop, the old deco stops have to be removed earlier as otherwise a new waypoint later than a previous generated
gas switch inherits the gas of the old switch.

Signed-off-by: Robert C. Helling <helling@atdotde.de>