Tissue saturation plot a la Sherwater Pretel
This adds a toolbox icon to turn on a tissue plot inspired by the bar graph of the Sherwater
Petrel,

It shows the inert gas partial pressures for individual compartments. If they are below
the ambient pressure (grey line) they are shown in units of the ambient pressure, if
they are above, the excess is shown as a percentage of the allowed overpressure for
plain Buehlmann. So it has the same units as a gradient factor. Thus also the a
gradient factor line (for the current depth) is shown.

The different tissues get different colors, greener for the faster ones and bluer
for the slower ones.

Positioning and on/off icon action still need some tender loving care.

Signed-off-by: Robert C. Helling <helling@atdotde.de>