planner: QMutex unlock when returning early
The plan() function is used also outside the planner, for example,
to construct a default profile when manually adding a dive. In
this case, the plan() function is returned early, and the planner
lock QMutex was not unlocked, causing dead lock.

Fixed here. When returning early, unlock the planner Qmutex.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>