Demo how to localize floating point printing
This is an example how one could localize the printing of floats
from C code (while keeping the translation in tact).

This localizes only one put_format, others would need to be
handled similarly.

NOTE: This is only a demo and leaks memory. The string returned
be float2string needs to be freeed!!!

Signed-off-by: Robert C. Helling <helling@atdotde.de>