Improve OTU calculations
Implement the protocol in Erik Baker's document
"Oxygen Toxicity Calculations". Different
formulae are used, depending on whether a dive
segment is constant depth or ascending/descending
and depending on the PO2 at the start and at the
end of the dive segment, assuming a constant
rate of descent or ascent. Provision is made for
PSCR and CCR dive logs and dive plans. In the
case of dive logs, the values of o2 sensors are
used if there are data from such sensors. For CCR
only the data from the first O2 sensor is used even if
there are more than one sensor. This is a potential
weakness, but this function is probably NOT the
place to calculate mean o2 values accross all sensors
and to emulate voting logic to reject info from
aberrant sensors.

Signed-off-by: willemferguson <willemferguson@zoology.up.ac.za>