Add some debugging info
The data follow in the planner is as follows:

From the info in the UI, a sturct dive plan is created.
This contains information about the available gases at
special entries at t=0.

This commit adds print statements where this list is constructed.

No cyliders not in use should be added, but it seems this field is filled wrong.

Signed-off-by: Robert C. Helling <helling@atdotde.de>